+++ trackpy/locate_functions/brightfield_ring.py
